Oh no, too many updates!

+++ server/src/controllers/productsController.js
@@ -3,6 +3,10 @@ const client = require('../config/client')
 const getAllProducts = async (req, res) => {
     try {
         // get all products where available is true
+        if (req.isAdmin) {
+            const { rows } = await client.query('SELECT * FROM product')
+            res.status(200).json(rows)
+        }
         const { rows } = await client.query('SELECT * FROM product WHERE available = true')
         res.status(200).json(rows)
     } catch (error) {
@@ -12,11 +16,12 @@ const getAllProducts = async (req, res) => {
 
 const createProduct = async (req, res) => {
     try {
+        if (!req.isAdmin) return res.status(401).json({ message: 'Unauthorized' })
         const { rows } = await client.query('INSERT INTO product (name, price) VALUES ($1, $2) RETURNING *', [
             req.body.name,
             req.body.price,
         ])
-        res.json(rows)
+        res.status(201).json(rows)
     } catch (error) {
         console.error(error)
     }
@@ -33,6 +38,7 @@ const getProductById = async (req, res) => {
 
 const updateProduct = async (req, res) => {
     try {
+        if (!req.isAdmin) return res.status(401).json({ message: 'Unauthorized' })
         const { rows } = await client.query('UPDATE product SET name = $1, price = $2 WHERE id = $3 RETURNING *', [
             req.body.name,
             req.body.price,
@@ -46,6 +52,7 @@ const updateProduct = async (req, res) => {
 
 const deleteProduct = async (req, res) => {
     try {
+        if (!req.isAdmin) return res.status(401).json({ message: 'Unauthorized' })
         const { rows } = await client.query('DELETE FROM product WHERE id = $1 RETURNING *', [req.params.id])
         res.json(rows)
     } catch (error) {
